fix(platform-cloudflare): isolate shared cluster handle types
